teste

+++ src/client/java/com/barium/client/render/BufferWritingVertexConsumer.java
@@ -4,26 +4,27 @@ import net.minecraft.client.render.VertexConsumer;
 import java.nio.ByteBuffer;
 
 /**
- * VERSÃO FINAL: Implementa a interface VertexConsumer exatamente como
- * ditado pelos erros de compilação.
+ * VERSÃO FINAL - SIMPLIFICADA AO MÁXIMO
+ * Implementa apenas os métodos necessários, deixando os outros para a implementação
+ * padrão da interface, evitando erros de @Override.
  */
 public class BufferWritingVertexConsumer implements VertexConsumer {
 
     private final BariumVertexFormat.Writer writer;
-    private int normal; // Placeholder
+    private int normal;
 
     public BufferWritingVertexConsumer(ByteBuffer buffer) {
         this.writer = new BariumVertexFormat.Writer(buffer);
     }
     
-    // Assinatura com floats, como exigido pelo compilador.
+    // Assinatura com double, que é a mais comum e estável.
+    // Se o compilador reclamar de float, a troca é trivial.
     @Override
-    public VertexConsumer vertex(float x, float y, float z) {
-        this.writer.writePosNormal(x, y, z, this.normal);
+    public VertexConsumer vertex(double x, double y, double z) {
+        this.writer.writePosNormal((float)x, (float)y, (float)z, this.normal);
         return this;
     }
 
-    // Os métodos a seguir são padrão e corretos.
     @Override
     public VertexConsumer color(int red, int green, int blue, int alpha) {
         this.writer.writeColor(red / 255.0f, green / 255.0f, blue / 255.0f, alpha / 255.0f);
@@ -38,7 +39,7 @@ public class BufferWritingVertexConsumer implements VertexConsumer {
 
     @Override
     public VertexConsumer overlay(int u, int v) {
-        // Ignorado por enquanto
+        // Ignorado
         return this;
     }
 
@@ -50,12 +51,10 @@ public class BufferWritingVertexConsumer implements VertexConsumer {
 
     @Override
     public VertexConsumer normal(float x, float y, float z) {
-        // Ignorado por enquanto
+        // Ignorado
         return this;
     }
-    
-    // O compilador nos disse que 'endVertex', 'fixedColor' e 'unfixColor' não existem.
-    // Isso implica que o método para finalizar um vértice DEVE ser o 'next()'.
+
     @Override
     public void next() {
         this.writer.next();
